Size is not grandeur, and territory does not make a nation

+++ 0x00-hello_world/6-size.c
@@ -11,11 +11,11 @@ int main(void)
 	long long int ll;
 	float f;
 
-	printf("Size of char: %lu byte(s)\n", (unsigned long)sizeof(c));
-	printf("Size of int: %lu byte(s)\n", (unsigned long)sizeof(i));
-	printf("Size of long int: %lu byte(s)\n", (unsigned long)sizeof(l));
-	printf("Size of long long int: %lu byte(s)\n", (unsigned long)sizeof(ll));
-	printf("Size of float: %lu byte(s)\n", (unsigned long)sizeof(f));
+	printf("Size of char: %lu byte(s)\n", sizeof(c));
+	printf("Size of int: %lu byte(s)\n", sizeof(i));
+	printf("Size of long int: %lu byte(s)\n", sizeof(l));
+	printf("Size of long long int: %lu byte(s)\n", sizeof(ll));
+	printf("Size of float: %lu byte(s)\n", sizeof(f));
 
 	return (0);
 }
